requirements.txt
Flask
SQLAlchemy
psycopg[binary]
psycopg2-binary
gunicorn
Alembic
python-dotenv
openpyxl
